+++ API/database/migrations/2024_02_12_092432_create_order_product_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreign('order_id')->references('id')->on('orders');
             $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products');
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 5);
         });
     }
